+++ assets/app.js
@@ -78,7 +78,7 @@ function navBar(active){
     <a class="brand" href="./"><span class="dot"></span>NYC Modernization</a>
     <a class="link" href="./#domains">Domains</a>
     <a class="link" href="./synthesis.html">Synthesis</a>
-    <a class="link" href="./#method">Method</a>
+    <a class="link" href="./docs.html?f=ROADMAP.md">Roadmap</a>
     <a class="link" href="./docs.html?f=CHANGELOG.md">Changelog</a>
     <span class="spacer"></span>
     <a class="link ghost" href="https://github.com/api-evangelist/nyc">GitHub</a>